+++ src/cli.rs
@@ -9,6 +9,9 @@ pub enum Cli {
     /// Print an example configuration file which can be piped into website-stalker.yaml
     ExampleConfig,
 
+    /// Print the JSON Schema for the configuration file
+    JsonSchema,
+
     /// Initialize the current directory with a git repository and a configuration file (website-stalker.yaml)
     #[command(hide = true)]
     Init,